Extract server manipulation to exec command
