+++ cmd/server.go
@@ -1,11 +1,21 @@
 package cmd
 
-import "github.com/spf13/cobra"
+import (
+	"github.com/spf13/cobra"
+
+	"github.com/traPtitech/DevOpsBot/pkg/config"
+	"github.com/traPtitech/DevOpsBot/pkg/server"
+)
 
 var serverCmd = &cobra.Command{
-	Use:   "server",
-	Short: "ConoHa server manipulation",
+	Use:          "server",
+	Short:        "ConoHa server manipulation",
+	SilenceUsage: true, // Do not display command usage when RunE returns error
 	RunE: func(cmd *cobra.Command, args []string) error {
-		panic("implement me")
+		s, err := server.Compile(&config.C.Commands.Servers)
+		if err != nil {
+			return err
+		}
+		return s.Execute(args)
 	},
 }
