+++ pkg/bot/help.go
@@ -13,7 +13,6 @@ func (h *HelpCommand) Execute(ctx *Context) error {
 	var lines []string
 	lines = append(lines, fmt.Sprintf("## DevOpsBot v%s", utils.Version()))
 	lines = append(lines, fmt.Sprintf("- `%sdeploy` - Do deployments", config.C.Prefix))
-	lines = append(lines, fmt.Sprintf("- `%sserver` - Server management", config.C.Prefix))
 	lines = append(lines, fmt.Sprintf("- `%shelp` - This help", config.C.Prefix))
 	return ctx.Reply(lines...)
 }
